refactor: ResultView 도메인 refactoring
