+++ src/main/java/step3/view/ResultView.java
@@ -1,7 +1,6 @@
 package step3.view;
 
-import step3.Car;
-import step3.Distance;
+import step3.domain.Car;
 
 import java.util.List;
 
@@ -21,8 +20,7 @@ public class ResultView {
     }
 
     public static void getRoundResultWithName(List<Car> cars, String sign) {
-        cars.stream()
-                .forEach(c -> printCountBySignWithName(sign, c.getName().getName(), c.getDistance().getMileage()));
+        cars.forEach(c -> printCountBySignWithName(sign, c.getName().getName(), c.getDistance().getMileage()));
         System.out.println();
     }
 
@@ -42,16 +40,9 @@ public class ResultView {
         System.out.println(name + " : "  + result);
     }
 
-    public static void printWinner(List<Car> cars) {
-        Distance distance = cars.stream()
-                    .map(c -> c.getDistance())
-                    .max((c1, c2) -> (c1.getMileage() > c2.getMileage()) ? 1 : (c1.getMileage() == c2.getMileage()) ? 0 : -1)
-                    .get();
-
-        String carNames = cars.stream()
-                .filter(c -> c.getDistance().equals(distance))
-                .map(c -> c.getName().getName())
-                .collect(joining(", "));
+    public static void printWinner(List<String> names, String sign) {
+        String carNames = names.stream()
+                .collect(joining(sign));
         System.out.println(carNames + "가 최종 우승했습니다.");
     }
 }
