+++ src/test/java/step3/view/ResultViewTest.java
@@ -1,7 +1,7 @@
 package step3.view;
 
-import step3.model.RacingCarGame;
-import step3.model.RandomMovingStrategy;
+import step3.domain.RacingCarGame;
+import step3.domain.strategy.RandomMovingStrategy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,7 +14,7 @@ public class ResultViewTest {
         RacingCarGame racingCarGame = new RacingCarGame(3);
 
         for (int i = 0; i < 5; i++) {
-            racingCarGame.play(new RandomMovingStrategy());
+            racingCarGame.play(new RandomMovingStrategy(3, 1));
             ResultView.getRoundResult(racingCarGame.getCars(), "-");
         }
     }
@@ -25,7 +25,7 @@ public class ResultViewTest {
         RacingCarGame racingCarGame = new RacingCarGame(new String[]{"BLACK", "WHITE", "BLUE"});
 
         for (int i = 0; i < 5; i++) {
-            racingCarGame.play(new RandomMovingStrategy());
+            racingCarGame.play(new RandomMovingStrategy(3, 1));
             ResultView.getRoundResultWithName(racingCarGame.getCars(), "-");
         }
     }
@@ -36,8 +36,8 @@ public class ResultViewTest {
         RacingCarGame racingCarGame = new RacingCarGame(new String[]{"BLACK", "WHITE", "BLUE"});
 
         for (int i = 0; i < 5; i++) {
-            racingCarGame.play(new RandomMovingStrategy());
+            racingCarGame.play(new RandomMovingStrategy(3, 1));
         }
-        ResultView.printWinner(racingCarGame.getCars());
+        ResultView.printWinner(racingCarGame.getWinnerCarNames(), ", ");
     }
 }
